ENH: add spacecraft methods
