+++ pysatMissionPlanning/instruments/pysat_sgp4.py
@@ -40,11 +40,12 @@ def init(self):
     import pysatMissionPlanning.methods.aacgmv2 as methaacgm
     import pysatMissionPlanning.methods.apexpy as methapex
     import pysatMissionPlanning.methods.pyglow as methglow
+    import pysatMissionPlanning.methods.spacecraft as methsc
 
     self.custom.add(methapex.add_quasi_dipole_coordinates, 'modify')
     self.custom.add(methaacgm.add_aacgm_coordinates, 'modify')
-    self.custom.add(calculate_ecef_velocity, 'modify')
-    self.custom.add(add_sc_attitude_vectors, 'modify')
+    self.custom.add(methsc.calculate_ecef_velocity, 'modify')
+    self.custom.add(methsc.add_sc_attitude_vectors, 'modify')
     # Thermal Ion Parameters
     self.custom.add(methglow.add_iri_thermal_plasma, 'modify')
     # Thermal Neutral parameters
@@ -57,9 +58,10 @@ def init(self):
     in_meta = {'desc': 'IGRF geomagnetic field expressed in the s/c basis.',
                'units': 'nT'}
     # project IGRF
-    self.custom.add(project_ecef_vector_onto_sc, 'modify', 'end', 'B_ecef_x',
-                    'B_ecef_y', 'B_ecef_z', 'B_sc_x', 'B_sc_y', 'B_sc_z',
-                    meta=[in_meta.copy(), in_meta.copy(), in_meta.copy()])
+    self.custom.add(methsc.project_ecef_vector_onto_sc, 'modify', 'end',
+                    'B_ecef_x', 'B_ecef_y', 'B_ecef_z', 'B_sc_x', 'B_sc_y',
+                    'B_sc_z', meta=[in_meta.copy(), in_meta.copy(),
+                                    in_meta.copy()])
     # project total wind vector
     self.custom.add(project_hwm_onto_sc, 'modify')
 
@@ -283,263 +285,23 @@ def download(date_array, tag, sat_id, data_path=None, user=None,
     pass
 
 
-def add_sc_attitude_vectors(inst):
-    """
-    Add attitude vectors for spacecraft assuming ram pointing.
-
-    Presumes spacecraft is pointed along the velocity vector (x), z is
-    generally nadir pointing (positive towards Earth), and y completes the
-    right handed system (generally southward).
-
-    Notes
-    -----
-        Expects velocity and position of spacecraft in Earth Centered
-        Earth Fixed (ECEF) coordinates to be in the instrument object
-        and named velocity_ecef_* (*=x,y,z) and position_ecef_* (*=x,y,z)
-
-        Adds attitude vectors for spacecraft in the ECEF basis by calculating
-        the scalar product of each attitude vector with each component of ECEF.
-
-    Parameters
-    ----------
-    inst : pysat.Instrument
-        Instrument object
-
-    Returns
-    -------
-    None
-        Modifies pysat.Instrument object in place to include S/C attitude
-        unit vectors, expressed in ECEF basis. Vectors are named
-        sc_(x,y,z)hat_ecef_(x,y,z).
-        sc_xhat_ecef_x is the spacecraft unit vector along x (positive along
-        velocity vector) reported in ECEF, ECEF x-component.
-
-    """
-    import pysatMagVect
-
-    # ram pointing is along velocity vector
-    inst['sc_xhat_ecef_x'], inst['sc_xhat_ecef_y'], inst['sc_xhat_ecef_z'] = \
-        pysatMagVect.normalize_vector(inst['velocity_ecef_x'],
-                                      inst['velocity_ecef_y'],
-                                      inst['velocity_ecef_z'])
-
-    # begin with z along Nadir (towards Earth)
-    # if orbit isn't perfectly circular, then the s/c z vector won't
-    # point exactly along nadir. However, nadir pointing is close enough
-    # to the true z (in the orbital plane) that we can use it to get y,
-    # and use x and y to get the real z
-    inst['sc_zhat_ecef_x'], inst['sc_zhat_ecef_y'], inst['sc_zhat_ecef_z'] = \
-        pysatMagVect.normalize_vector(-inst['position_ecef_x'],
-                                      -inst['position_ecef_y'],
-                                      -inst['position_ecef_z'])
-
-    # get y vector assuming right hand rule
-    # Z x X = Y
-    inst['sc_yhat_ecef_x'], inst['sc_yhat_ecef_y'], inst['sc_yhat_ecef_z'] = \
-        pysatMagVect.cross_product(inst['sc_zhat_ecef_x'],
-                                   inst['sc_zhat_ecef_y'],
-                                   inst['sc_zhat_ecef_z'],
-                                   inst['sc_xhat_ecef_x'],
-                                   inst['sc_xhat_ecef_y'],
-                                   inst['sc_xhat_ecef_z'])
-    # normalize since Xhat and Zhat from above may not be orthogonal
-    inst['sc_yhat_ecef_x'], inst['sc_yhat_ecef_y'], inst['sc_yhat_ecef_z'] = \
-        pysatMagVect.normalize_vector(inst['sc_yhat_ecef_x'],
-                                      inst['sc_yhat_ecef_y'],
-                                      inst['sc_yhat_ecef_z'])
-
-    # strictly, need to recalculate Zhat so that it is consistent with RHS
-    # just created
-    # Z = X x Y
-    inst['sc_zhat_ecef_x'], inst['sc_zhat_ecef_y'], inst['sc_zhat_ecef_z'] = \
-        pysatMagVect.cross_product(inst['sc_xhat_ecef_x'],
-                                   inst['sc_xhat_ecef_y'],
-                                   inst['sc_xhat_ecef_z'],
-                                   inst['sc_yhat_ecef_x'],
-                                   inst['sc_yhat_ecef_y'],
-                                   inst['sc_yhat_ecef_z'])
-
-    # Adding metadata
-    inst.meta['sc_xhat_ecef_x'] = {'units': '',
-                                   'desc': 'S/C attitude (x-direction, ram) ' +
-                                   'unit vector, expressed in ECEF basis, ' +
-                                   'x-component'}
-    inst.meta['sc_xhat_ecef_y'] = {'units': '',
-                                   'desc': 'S/C attitude (x-direction, ram) ' +
-                                   'unit vector, expressed in ECEF basis, ' +
-                                   'y-component'}
-    inst.meta['sc_xhat_ecef_z'] = {'units': '',
-                                   'desc': 'S/C attitude (x-direction, ram) ' +
-                                   'unit vector, expressed in ECEF basis, ' +
-                                   'z-component'}
-
-    inst.meta['sc_zhat_ecef_x'] = {'units': '',
-                                   'desc': 'S/C attitude (z-direction, ' +
-                                   'generally nadir) unit vector, expressed ' +
-                                   'in ECEF basis, x-component'}
-    inst.meta['sc_zhat_ecef_y'] = {'units': '',
-                                   'desc': 'S/C attitude (z-direction, ' +
-                                   'generally nadir) unit vector, expressed ' +
-                                   'in ECEF basis, y-component'}
-    inst.meta['sc_zhat_ecef_z'] = {'units': '',
-                                   'desc': 'S/C attitude (z-direction, ' +
-                                   'generally nadir) unit vector, expressed ' +
-                                   'in ECEF basis, z-component'}
-
-    inst.meta['sc_yhat_ecef_x'] = {'units': '',
-                                   'desc': 'S/C attitude (y-direction, ' +
-                                   'generally south) unit vector, expressed ' +
-                                   'in ECEF basis, x-component'}
-    inst.meta['sc_yhat_ecef_y'] = {'units': '',
-                                   'desc': 'S/C attitude (y-direction, ' +
-                                   'generally south) unit vector, expressed ' +
-                                   'in ECEF basis, y-component'}
-    inst.meta['sc_yhat_ecef_z'] = {'units': '',
-                                   'desc': 'S/C attitude (y-direction, ' +
-                                   'generally south) unit vector, expressed ' +
-                                   'in ECEF basis, z-component'}
-
-    # check what magnitudes we get
-    mag = np.sqrt(inst['sc_zhat_ecef_x']**2 + inst['sc_zhat_ecef_y']**2 +
-                  inst['sc_zhat_ecef_z']**2)
-    idx, = np.where((mag < .999999999) | (mag > 1.000000001))
-    if len(idx) > 0:
-        print(mag[idx])
-        raise RuntimeError('Unit vector generation failure. Not sufficently ' +
-                           'orthogonal.')
-
-    return
-
-
-def calculate_ecef_velocity(inst):
-    """
-    Calculates spacecraft velocity in ECEF frame.
-
-    Presumes that the spacecraft velocity in ECEF is in
-    the input instrument object as position_ecef_*. Uses a symmetric
-    difference to calculate the velocity thus endpoints will be
-    set to NaN. Routine should be run using pysat data padding feature
-    to create valid end points.
-
-    Parameters
-    ----------
-    inst : pysat.Instrument
-        Instrument object
-
-    Returns
-    -------
-    None
-        Modifies pysat.Instrument object in place to include ECEF velocity
-        using naming scheme velocity_ecef_* (*=x,y,z)
-
-    """
-
-    x = inst['position_ecef_x']
-    vel_x = (x.values[2:] - x.values[0:-2])/2.
-
-    y = inst['position_ecef_y']
-    vel_y = (y.values[2:] - y.values[0:-2])/2.
-
-    z = inst['position_ecef_z']
-    vel_z = (z.values[2:] - z.values[0:-2])/2.
-
-    inst[1:-1, 'velocity_ecef_x'] = vel_x
-    inst[1:-1, 'velocity_ecef_y'] = vel_y
-    inst[1:-1, 'velocity_ecef_z'] = vel_z
-
-    inst.meta['velocity_ecef_x'] = {'units': 'km/s',
-                                    'desc': 'Velocity of satellite ' +
-                                    'calculated with respect to ECEF frame.'}
-    inst.meta['velocity_ecef_y'] = {'units': 'km/s',
-                                    'desc': 'Velocity of satellite ' +
-                                    'calculated with respect to ECEF frame.'}
-    inst.meta['velocity_ecef_z'] = {'units': 'km/s',
-                                    'desc': 'Velocity of satellite ' +
-                                    'calculated with respect to ECEF frame.'}
-    return
-
-
-def project_ecef_vector_onto_sc(inst, x_label, y_label, z_label,
-                                new_x_label, new_y_label, new_z_label,
-                                meta=None):
-    """Express input vector using s/c attitude directions
-
-    x - ram pointing
-    y - generally southward
-    z - generally nadir
-
-    Parameters
-    ----------
-    x_label : string
-        Label used to get ECEF-X component of vector to be projected
-    y_label : string
-        Label used to get ECEF-Y component of vector to be projected
-    z_label : string
-        Label used to get ECEF-Z component of vector to be projected
-    new_x_label : string
-        Label used to set X component of projected vector
-    new_y_label : string
-        Label used to set Y component of projected vector
-    new_z_label : string
-        Label used to set Z component of projected vector
-    meta : array_like of dicts (None)
-        Dicts contain metadata to be assigned.
-    """
-
-    import pysatMagVect
-
-    x, y, z = \
-        pysatMagVect.project_ecef_vector_onto_basis(inst[x_label],
-                                                    inst[y_label],
-                                                    inst[z_label],
-                                                    inst['sc_xhat_ecef_x'],
-                                                    inst['sc_xhat_ecef_y'],
-                                                    inst['sc_xhat_ecef_z'],
-                                                    inst['sc_yhat_ecef_x'],
-                                                    inst['sc_yhat_ecef_y'],
-                                                    inst['sc_yhat_ecef_z'],
-                                                    inst['sc_zhat_ecef_x'],
-                                                    inst['sc_zhat_ecef_y'],
-                                                    inst['sc_zhat_ecef_z'])
-    inst[new_x_label] = x
-    inst[new_y_label] = y
-    inst[new_z_label] = z
-
-    if meta is not None:
-        inst.meta[new_x_label] = meta[0]
-        inst.meta[new_y_label] = meta[1]
-        inst.meta[new_z_label] = meta[2]
-
-    return
-
-
 def project_hwm_onto_sc(inst):
 
-    import pysatMagVect
+    import pysatMissionPlanning.methods.attitude as methatt
 
-    total_wind_x = inst['zonal_wind']*inst['unit_zonal_wind_ecef_x'] + \
-        inst['meridional_wind']*inst['unit_mer_wind_ecef_x']
-    total_wind_y = inst['zonal_wind']*inst['unit_zonal_wind_ecef_y'] + \
-        inst['meridional_wind']*inst['unit_mer_wind_ecef_y']
-    total_wind_z = inst['zonal_wind']*inst['unit_zonal_wind_ecef_z'] + \
-        inst['meridional_wind']*inst['unit_mer_wind_ecef_z']
+    inst['total_wind_x'] = \
+        (inst['zonal_wind']*inst['unit_zonal_wind_ecef_x'] +
+         inst['meridional_wind']*inst['unit_mer_wind_ecef_x'])
+    inst['total_wind_y'] = \
+        (inst['zonal_wind']*inst['unit_zonal_wind_ecef_y'] +
+         inst['meridional_wind']*inst['unit_mer_wind_ecef_y'])
+    inst['total_wind_z'] = \
+        (inst['zonal_wind']*inst['unit_zonal_wind_ecef_z'] +
+         inst['meridional_wind']*inst['unit_mer_wind_ecef_z'])
 
-    x, y, z = \
-        pysatMagVect.project_ecef_vector_onto_basis(total_wind_x,
-                                                    total_wind_y,
-                                                    total_wind_z,
-                                                    inst['sc_xhat_ecef_x'],
-                                                    inst['sc_xhat_ecef_y'],
-                                                    inst['sc_xhat_ecef_z'],
-                                                    inst['sc_yhat_ecef_x'],
-                                                    inst['sc_yhat_ecef_y'],
-                                                    inst['sc_yhat_ecef_z'],
-                                                    inst['sc_zhat_ecef_x'],
-                                                    inst['sc_zhat_ecef_y'],
-                                                    inst['sc_zhat_ecef_z'])
-    inst['sim_wind_sc_x'] = x
-    inst['sim_wind_sc_y'] = y
-    inst['sim_wind_sc_z'] = z
+    methatt.project_ecef_vector_onto_sc(inst, 'total_wind_x', 'total_wind_y',
+                                        'total_wind_z', 'sim_wind_sc_x',
+                                        'sim_wind_sc_y', 'sim_wind_sc_z')
 
     inst.meta['sim_wind_sc_x'] = {'units': 'm/s',
                                   'long_name': 'Simulated x-vector ' +
