+++ pysatMissionPlanning/methods/__init__.py
@@ -3,6 +3,6 @@ pysatMissionPlanning.methods is a module that provides
 the methods to interface with numerous empirical model packages
 """
 
-from pysatMissionPlanning.methods import aacgmv2, apexpy, pyglow
+from pysatMissionPlanning.methods import aacgmv2, apexpy, pyglow, spacecraft
 
-__all__ = ['aacgmv2', 'apexpy', 'pyglow']
+__all__ = ['aacgmv2', 'apexpy', 'pyglow', 'spacecraft']
